refactor combine shared context logic
